requirements.txt
PyYAML
torch
torchvision
pytorchvideo
wandb
pandas
numpy
scikit-image
scikit-learn
stam-pytorch
timesformer-pytorch
opencv-python-headless
zarr
matplotlib
seaborn
torchmetrics
git+https://github.com/HeartWise-AI/HeartWise_StatPlots.git@dev